Final refactor and 100% test coverage again
